+++ DesignPatterns.SourceGenerators/Generators/EquatableArray.cs
@@ -18,20 +18,23 @@ internal readonly struct EquatableArray<T> : IEquatable<EquatableArray<T>>, IRea
         _array = array ?? Array.Empty<T>();
     }
 
-    public int Count => _array.Length;
+    public int Count => _array?.Length ?? 0;
 
-    public T this[int index] => _array[index];
+    public T this[int index] => _array![index];
 
     public bool Equals(EquatableArray<T> other)
     {
-        if (_array.Length != other._array.Length)
+        var left = _array ?? Array.Empty<T>();
+        var right = other._array ?? Array.Empty<T>();
+
+        if (left.Length != right.Length)
         {
             return false;
         }
 
-        for (var i = 0; i < _array.Length; i++)
+        for (var i = 0; i < left.Length; i++)
         {
-            if (!_array[i].Equals(other._array[i]))
+            if (!left[i].Equals(right[i]))
             {
                 return false;
             }
@@ -46,7 +49,7 @@ internal readonly struct EquatableArray<T> : IEquatable<EquatableArray<T>>, IRea
     {
         // Manual hash combine to avoid HashCode (not available on netstandard2.0).
         var hash = 0;
-        foreach (var item in _array)
+        foreach (var item in _array ?? Array.Empty<T>())
         {
             hash = (hash * 31) + (item?.GetHashCode() ?? 0);
         }
@@ -58,11 +61,11 @@ internal readonly struct EquatableArray<T> : IEquatable<EquatableArray<T>>, IRea
 
     public static bool operator !=(EquatableArray<T> left, EquatableArray<T> right) => !left.Equals(right);
 
-    public Enumerator GetEnumerator() => new Enumerator(_array);
+    public Enumerator GetEnumerator() => new Enumerator(_array ?? Array.Empty<T>());
 
-    IEnumerator<T> IEnumerable<T>.GetEnumerator() => ((IEnumerable<T>)_array).GetEnumerator();
+    IEnumerator<T> IEnumerable<T>.GetEnumerator() => ((IEnumerable<T>)(_array ?? Array.Empty<T>())).GetEnumerator();
 
-    IEnumerator IEnumerable.GetEnumerator() => _array.GetEnumerator();
+    IEnumerator IEnumerable.GetEnumerator() => (_array ?? Array.Empty<T>()).GetEnumerator();
 
     public struct Enumerator
     {